+++ packages/homeassistant/src/index.ts
@@ -11,3 +11,5 @@ export * from "./HomeAssistantEntityStateTransport";
 export * from "./HomeAssistantEntityStatusPanelBinding";
 export * from "./HomeAssistantWebSocketProtocol";
 export * from "./HomeAssistantWebSocketTransport";
+export * from "./HomeAssistantBrowserWebSocket";
+export * from "./HomeAssistantRuntimeConnection";